Environment: Fix ArrayAccess statement in function block using argument
Previously if you tried to index an array that was passed as an argument to a function, you would get an UnknownIdentifier thrown at you. This was due to 'arrayResolve' only looking up the in the array block and not querying the vars-block which the argument IS mapped to. Now 'arrayAccess' will call env.varResolve after searching the array block and finding nothing.