Interpreter & AST: Add 'Call' Expression node for globally dispatching calls
Call will now dispatch all call type during the parse phase. This is because to know what call we want to lookup, we need to be in execution/evaluation context. This also will add a 'more' member to 'BadCall'.

Parse: Add ClassDeclaration method for parsing classes. New 'argsExpr' method for getting a list of expressions (cleaned up from FunctionDeclaration).